Updated Change Non-Rating Issue to exist in it's own file and removed it from war_room.rb file based on feedback
